+++ app/src/app.js
@@ -86,13 +86,9 @@ async function init() {
             app.use(koaSimpleHealthCheck());
 
             app.use(RWAPIMicroservice.bootstrap({
-                name: config.get('service.name'),
-                info: require('../microservice/register.json'),
-                swagger: require('../microservice/public-swagger.json'),
                 logger,
-                baseURL: process.env.CT_URL,
-                url: process.env.LOCAL_URL,
-                token: process.env.CT_TOKEN,
+                gatewayURL: process.env.GATEWAY_URL,
+                microserviceToken: process.env.MICROSERVICE_TOKEN,
                 fastlyEnabled: process.env.FASTLY_ENABLED,
                 fastlyServiceId: process.env.FASTLY_SERVICEID,
                 fastlyAPIKey: process.env.FASTLY_APIKEY
@@ -103,14 +99,6 @@ async function init() {
             loader.loadRoutes(app);
 
             const server = app.listen(process.env.PORT, () => {
-                if (process.env.CT_REGISTER_MODE === 'auto') {
-                    RWAPIMicroservice.register().then(() => {
-                        logger.info('CT registration process started');
-                    }, (error) => {
-                        logger.error(error);
-                        process.exit(1);
-                    });
-                }
             });
 
             logger.info('Server started in ', process.env.PORT);