+++ app/test/e2e/utils/mocks.js
@@ -2,7 +2,7 @@ const nock = require('nock');
 const intersection = require('lodash/intersection');
 
 const createMockUser = (users) => {
-    nock(process.env.CT_URL)
+    nock(process.env.GATEWAY_URL)
         .post(
             '/auth/user/find-by-ids',
             (body) => intersection(body.ids, users.map((e) => e.id.toString())).length === body.ids.length
@@ -10,7 +10,7 @@ const createMockUser = (users) => {
         .reply(200, { data: users });
 };
 
-const createMockUserRole = (role, userID) => nock(process.env.CT_URL)
+const createMockUserRole = (role, userID) => nock(process.env.GATEWAY_URL)
     .get(`/auth/user/ids/${role}`)
     .reply(200, { data: [userID] });