+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 16/06/2021
+
+- Update `rw-api-microservice-node` to remove CT integration support.
+
 ## 09/06/2021
 
 - Add support for hosts from `x-rw-domain` header when generating pagination links.